Signup and signin user done and authenticated all query and mutation with cognito user pool

+++ Task-2 - User and Blogs/Blogs/blogsHandler.js
@@ -1,25 +1,30 @@
 'use strict';
 const AWS = require('aws-sdk');
 const uuid = require('node-uuid');
+const {auth} = require('../Authorization/auth')
+
 
 const dynamoDb = new AWS.DynamoDB.DocumentClient();
 
 const createBlog = async (event) => {
   try {
-    const id = uuid.v4();
-    const params = {
-      TableName: process.env.BlogsTable,
-      Item :{
-        id,
-        ...event.arguments
-      }
-    };
-    const blog = await dynamoDb.put(params).promise();
-    
-    return {
+    const authToken = event.request.headers.authorization.split(' ')[1];
+    if(auth(authToken)){
+      const id = uuid.v4();
+      const params = {
+        TableName: process.env.BlogsTable,
+        Item :{
           id,
           ...event.arguments
         }
+      };
+      const blog = await dynamoDb.put(params).promise();
+      return {
+        id,
+        ...event.arguments
+      }
+    }
+    
   } catch (error) {
     return {
       message : error.message
@@ -29,12 +34,13 @@ const createBlog = async (event) => {
 
 const updateBlog = async (event) => {
   try {
+    const authToken = event.request.headers.authorization.split(' ')[1];
+    if(auth(authToken)){
       const {id,userId, title, content, visibility} = event.arguments
-      var params = {
+      const params = {
         TableName : process.env.BlogsTable,
         Key: { 
-            id,
-            userId 
+            id 
         },
         UpdateExpression : 'set #a = :title, #b = :content, #c = :visibility',
         ExpressionAttributeNames: { '#a' : 'title', '#b': 'content', '#c' : 'visibility' },
@@ -43,7 +49,9 @@ const updateBlog = async (event) => {
       const blogUpdate = await dynamoDb.update(params).promise();
       return {
         ...event.arguments
+      }
     }
+      
   } catch (error) {
       console.log(error)
       return {
@@ -54,18 +62,21 @@ const updateBlog = async (event) => {
 
 const deleteBlog = async (event) => {
   try {
+    const authToken = event.request.headers.authorization.split(' ')[1];
+    if(auth(authToken)){
       const {id , userId} = event.arguments
-    const params = {
-      TableName: process.env.BlogsTable,
-      Key: {
-        id,
-        userId
-      },
-    };
-    const blogDelete = await dynamoDb.delete(params).promise();
-    return {
-      id : event.arguments.id
+      const params = {
+        TableName: process.env.BlogsTable,
+        Key: {
+          id
+        },
+      };
+      const blogDelete = await dynamoDb.delete(params).promise();
+      return {
+        id : event.arguments.id
+      }
     }
+      
   } catch (error) {
       return {
         message: error.message
@@ -75,18 +86,22 @@ const deleteBlog = async (event) => {
 
 const getBlogById = async (event) => {
   try {
-    const {id} = event.arguments
-    const params = {
-      TableName: process.env.BlogsTable,
-      Key: {
-        id
-      },
-    };
+    const authToken = event.request.headers.authorization.split(' ')[1];
+    if(auth(authToken)){
+      const {id} = event.arguments
+      const params = {
+        TableName: process.env.BlogsTable,
+        Key: {
+          id
+        },
+      };
 
-    const blog = await dynamoDb.get(params).promise();
-    return {
-      ...blog.Item
+      const blog = await dynamoDb.get(params).promise();
+      return {
+        ...blog.Item
+      }
     }
+    
   } catch (error) {
       return {
         message: error.message
@@ -96,18 +111,22 @@ const getBlogById = async (event) => {
 
 const getBlogs = async (event) => {
   try {
-    const {id,userId} = event.arguments
-    const params = {
-        TableName: process.env.BlogsTable,
-        IndexName : "UserIdIndex",
-        KeyConditionExpression: 'userId = :userId',
-        ExpressionAttributeValues: {
-            ':userId': userId
-        }
-    };
+    const authToken = event.request.headers.authorization.split(' ')[1];
+    if(auth(authToken)){
+      const {id,userId} = event.arguments
+      const params = {
+          TableName: process.env.BlogsTable,
+          IndexName : "UserIdIndex",
+          KeyConditionExpression: 'userId = :userId',
+          ExpressionAttributeValues: {
+              ':userId': userId
+          }
+      };
 
-    const blog = await dynamoDb.query(params).promise();
-    return blog.Items
+      const blog = await dynamoDb.query(params).promise();
+      return blog.Items
+    }
+    
   } catch (error) {
       console.log(error);
       return {
